+++ volume-cartographer/apps/VC3D/ViewerManager.cpp
@@ -16,8 +16,88 @@
 #include <QMdiArea>
 #include <QMdiSubWindow>
 #include <QSettings>
+#include <QtConcurrent/QtConcurrent>
 #include <algorithm>
 #include <cmath>
+#include <optional>
+#include <nlohmann/json.hpp>
+#include <opencv2/core.hpp>
+
+namespace {
+struct CellRegion {
+    int rowStart = 0;
+    int rowEnd = 0;
+    int colStart = 0;
+    int colEnd = 0;
+};
+
+struct DirtyBoundsInfo {
+    cv::Rect rect;
+    int version = 0;
+};
+
+std::optional<DirtyBoundsInfo> readDirtyBounds(QuadSurface* surface)
+{
+    if (!surface || !surface->meta || !surface->meta->is_object()) {
+        return std::nullopt;
+    }
+
+    nlohmann::json& meta = *surface->meta;
+    auto it = meta.find("dirty_bounds");
+    if (it == meta.end() || !it->is_object()) {
+        return std::nullopt;
+    }
+
+    const int rowStart = it->value("row_start", -1);
+    const int rowEnd = it->value("row_end", -1);
+    const int colStart = it->value("col_start", -1);
+    const int colEnd = it->value("col_end", -1);
+
+    if (rowStart < 0 || colStart < 0 || rowEnd <= rowStart || colEnd <= colStart) {
+        return std::nullopt;
+    }
+
+    int version = it->value("version", 0);
+    if (version <= 0) {
+        auto versionIt = meta.find("dirty_bounds_version");
+        if (versionIt != meta.end() && versionIt->is_number_integer()) {
+            version = versionIt->get<int>();
+        }
+    }
+
+    DirtyBoundsInfo info;
+    info.rect = cv::Rect(colStart, rowStart, colEnd - colStart, rowEnd - rowStart);
+    info.version = std::max(version, 1);
+    return info;
+}
+
+std::optional<CellRegion> vertexRectToCellRegion(const cv::Rect& vertexRect,
+                                                 QuadSurface* surface)
+{
+    if (!surface) {
+        return std::nullopt;
+    }
+    const cv::Mat_<cv::Vec3f>* points = surface->rawPointsPtr();
+    if (!points || points->rows < 2 || points->cols < 2) {
+        return std::nullopt;
+    }
+
+    const int cellRowCount = points->rows - 1;
+    const int cellColCount = points->cols - 1;
+
+    CellRegion region;
+    region.rowStart = std::max(0, vertexRect.y - 1);
+    region.rowEnd = std::min(cellRowCount, vertexRect.y + vertexRect.height);
+    region.colStart = std::max(0, vertexRect.x - 1);
+    region.colEnd = std::min(cellColCount, vertexRect.x + vertexRect.width);
+
+    if (region.rowStart >= region.rowEnd || region.colStart >= region.colEnd) {
+        return std::nullopt;
+    }
+
+    return region;
+}
+} // namespace
 
 ViewerManager::ViewerManager(CSurfaceCollection* surfaces,
                              VCCollection* points,
@@ -43,6 +123,20 @@ ViewerManager::ViewerManager(CSurfaceCollection* surfaces,
     _surfacePatchSamplingStride = std::max(1, storedSampling);
     const float storedThickness = settings.value("viewer/intersection_thickness", 0.0f).toFloat();
     _intersectionThickness = std::max(0.0f, storedThickness);
+
+    _surfacePatchIndexWatcher =
+        new QFutureWatcher<std::shared_ptr<SurfacePatchIndex>>(this);
+    connect(_surfacePatchIndexWatcher,
+            &QFutureWatcher<std::shared_ptr<SurfacePatchIndex>>::finished,
+            this,
+            &ViewerManager::handleSurfacePatchIndexPrimeFinished);
+
+    if (_surfaces) {
+        connect(_surfaces,
+                &CSurfaceCollection::sendSurfaceChanged,
+                this,
+                &ViewerManager::handleSurfaceChanged);
+    }
 }
 
 CVolumeViewer* ViewerManager::createViewer(const std::string& surfaceName,
@@ -53,7 +147,7 @@ CVolumeViewer* ViewerManager::createViewer(const std::string& surfaceName,
         return nullptr;
     }
 
-    auto* viewer = new CVolumeViewer(_surfaces, mdiArea);
+    auto* viewer = new CVolumeViewer(_surfaces, this, mdiArea);
     auto* win = mdiArea->addSubWindow(viewer);
     win->setWindowTitle(title);
     win->setWindowFlags(Qt::WindowTitleHint | Qt::WindowMinMaxButtonsHint);
@@ -350,6 +444,11 @@ void ViewerManager::setSurfacePatchSamplingStride(int stride)
     QSettings settings(vc3d::settingsFilePath(), QSettings::IniFormat);
     settings.setValue("viewer/intersection_sampling_stride", _surfacePatchSamplingStride);
 
+    if (_surfacePatchIndex.setSamplingStride(_surfacePatchSamplingStride)) {
+        _surfacePatchIndexDirty = true;
+        _indexedSurfaces.clear();
+    }
+
     for (auto* viewer : _viewers) {
         if (viewer) {
             viewer->setSurfacePatchSamplingStride(_surfacePatchSamplingStride);
@@ -357,8 +456,23 @@ void ViewerManager::setSurfacePatchSamplingStride(int stride)
     }
 }
 
+SurfacePatchIndex* ViewerManager::surfacePatchIndex()
+{
+    rebuildSurfacePatchIndexIfNeeded();
+    if (_surfacePatchIndex.empty()) {
+        return nullptr;
+    }
+    return &_surfacePatchIndex;
+}
+
 void ViewerManager::primeSurfacePatchIndicesAsync()
 {
+    if (!_surfacePatchIndexWatcher) {
+        return;
+    }
+    if (_surfacePatchIndexWatcher->isRunning()) {
+        _surfacePatchIndexWatcher->waitForFinished();
+    }
     if (!_surfaces) {
         return;
     }
@@ -370,10 +484,161 @@ void ViewerManager::primeSurfacePatchIndicesAsync()
             quadSurfaces.push_back(quad);
         }
     }
-    for (auto* viewer : _viewers) {
-        if (viewer) {
-            viewer->primeSurfacePatchIndexAsync(quadSurfaces);
+    _pendingSurfacePatchIndexSurfaces = quadSurfaces;
+    if (_pendingSurfacePatchIndexSurfaces.empty()) {
+        _surfacePatchIndex.clear();
+        _indexedSurfaces.clear();
+        _surfacePatchIndexDirty = false;
+        return;
+    }
+
+    auto surfacesForTask = _pendingSurfacePatchIndexSurfaces;
+    auto future = QtConcurrent::run([surfacesForTask]() mutable -> std::shared_ptr<SurfacePatchIndex> {
+        auto index = std::make_shared<SurfacePatchIndex>();
+        index->rebuild(surfacesForTask);
+        return index;
+    });
+    _surfacePatchIndexWatcher->setFuture(future);
+}
+
+void ViewerManager::rebuildSurfacePatchIndexIfNeeded()
+{
+    if (!_surfacePatchIndexDirty) {
+        return;
+    }
+    _surfacePatchIndexDirty = false;
+
+    if (!_surfaces) {
+        _surfacePatchIndex.clear();
+        _indexedSurfaces.clear();
+        return;
+    }
+
+    std::vector<QuadSurface*> surfaces;
+    for (Surface* surf : _surfaces->surfaces()) {
+        if (auto* quad = dynamic_cast<QuadSurface*>(surf)) {
+            surfaces.push_back(quad);
         }
+    }
+
+    if (surfaces.empty()) {
+        _surfacePatchIndex.clear();
+        _indexedSurfaces.clear();
+        return;
+    }
+
+    _surfacePatchIndex.rebuild(surfaces);
+    _indexedSurfaces.clear();
+    _indexedSurfaces.insert(surfaces.begin(), surfaces.end());
+}
+
+void ViewerManager::handleSurfacePatchIndexPrimeFinished()
+{
+    if (!_surfacePatchIndexWatcher) {
+        return;
+    }
+    auto result = _surfacePatchIndexWatcher->future().result();
+    if (!result) {
+        _pendingSurfacePatchIndexSurfaces.clear();
+        return;
+    }
+    _surfacePatchIndex = std::move(*result);
+    _surfacePatchIndexDirty = false;
+    _indexedSurfaces.clear();
+    _indexedSurfaces.insert(_pendingSurfacePatchIndexSurfaces.begin(),
+                            _pendingSurfacePatchIndexSurfaces.end());
+    _pendingSurfacePatchIndexSurfaces.clear();
+}
+
+void ViewerManager::handleSurfaceChanged(std::string /*name*/, Surface* surf)
+{
+    bool affectsSurfaceIndex = false;
+    bool regionUpdated = false;
+    bool indexUpdated = false;
+
+    if (auto* quad = dynamic_cast<QuadSurface*>(surf)) {
+        affectsSurfaceIndex = true;
+        std::optional<cv::Rect> dirtyVertices;
+        bool dirtyBoundsRegressed = false;
+        const bool alreadyIndexed = _indexedSurfaces.count(quad) != 0;
+        if (auto dirtyInfo = readDirtyBounds(quad)) {
+            int lastVersion = 0;
+            auto it = _surfaceDirtyBoundsVersions.find(quad);
+            if (it != _surfaceDirtyBoundsVersions.end()) {
+                lastVersion = it->second;
+            }
+            if (dirtyInfo->version > lastVersion) {
+                dirtyVertices = dirtyInfo->rect;
+            } else if (dirtyInfo->version < lastVersion) {
+                dirtyBoundsRegressed = true;
+            }
+            _surfaceDirtyBoundsVersions[quad] = dirtyInfo->version;
+        } else {
+            _surfaceDirtyBoundsVersions.erase(quad);
+        }
+
+        bool skippedDueToExistingIndex = false;
+        if (!_surfacePatchIndexDirty) {
+            if (dirtyVertices) {
+                if (auto cellRegion = vertexRectToCellRegion(*dirtyVertices, quad)) {
+                    regionUpdated = _surfacePatchIndex.updateSurfaceRegion(
+                        quad,
+                        cellRegion->rowStart,
+                        cellRegion->rowEnd,
+                        cellRegion->colStart,
+                        cellRegion->colEnd);
+                }
+            }
+            if (!regionUpdated && (!alreadyIndexed || dirtyBoundsRegressed)) {
+                indexUpdated = _surfacePatchIndex.updateSurface(quad);
+            } else if (!regionUpdated && alreadyIndexed && !dirtyBoundsRegressed) {
+                skippedDueToExistingIndex = true;
+            }
+        }
+        if (dirtyBoundsRegressed) {
+            _surfacePatchIndexDirty = true;
+            _indexedSurfaces.erase(quad);
+        }
+        if (skippedDueToExistingIndex) {
+            regionUpdated = true;
+        }
+        if (regionUpdated || indexUpdated) {
+            _indexedSurfaces.insert(quad);
+        }
+    } else if (!surf) {
+        affectsSurfaceIndex = true;
+        if (_surfaces) {
+            std::unordered_set<const QuadSurface*> liveSurfaces;
+            auto surfaces = _surfaces->surfaces();
+            liveSurfaces.reserve(surfaces.size());
+            for (Surface* candidate : surfaces) {
+                if (auto* quadSurface = dynamic_cast<QuadSurface*>(candidate)) {
+                    liveSurfaces.insert(quadSurface);
+                }
+            }
+            for (auto it = _surfaceDirtyBoundsVersions.begin();
+                 it != _surfaceDirtyBoundsVersions.end();) {
+                if (!liveSurfaces.count(it->first)) {
+                    it = _surfaceDirtyBoundsVersions.erase(it);
+                } else {
+                    ++it;
+                }
+            }
+            for (auto it = _indexedSurfaces.begin(); it != _indexedSurfaces.end();) {
+                if (!liveSurfaces.count(*it)) {
+                    it = _indexedSurfaces.erase(it);
+                } else {
+                    ++it;
+                }
+            }
+        } else {
+            _surfaceDirtyBoundsVersions.clear();
+            _indexedSurfaces.clear();
+        }
+    }
+
+    if (affectsSurfaceIndex) {
+        _surfacePatchIndexDirty = _surfacePatchIndexDirty || !(regionUpdated || indexUpdated);
     }
 }
