+++ volume-cartographer/apps/VC3D/ViewerManager.hpp
@@ -2,12 +2,16 @@
 
 #include <QObject>
 #include <QString>
+#include <QFutureWatcher>
 
 #include <functional>
 #include <memory>
 #include <string>
 #include <unordered_map>
+#include <unordered_set>
 #include <vector>
+
+#include "vc/core/util/SurfacePatchIndex.hpp"
 
 class QMdiArea;
 class CVolumeViewer;
@@ -22,6 +26,8 @@ class VolumeOverlayController;
 class ChunkCache;
 class SegmentationModule;
 class Volume;
+class Surface;
+class QuadSurface;
 
 class ViewerManager : public QObject
 {
@@ -85,12 +91,17 @@ public:
     void setIntersectionThickness(float thickness);
     float intersectionThickness() const { return _intersectionThickness; }
     void setHighlightedSurfaceIds(const std::vector<std::string>& ids);
+    SurfacePatchIndex* surfacePatchIndex();
 
 signals:
     void viewerCreated(CVolumeViewer* viewer);
     void overlayWindowChanged(float low, float high);
     void volumeWindowChanged(float low, float high);
     void overlayVolumeAvailabilityChanged(bool hasOverlay);
+
+private slots:
+    void handleSurfacePatchIndexPrimeFinished();
+    void handleSurfaceChanged(std::string name, Surface* surf);
 
 private:
     CSurfaceCollection* _surfaces;
@@ -119,4 +130,12 @@ private:
     int _surfacePatchSamplingStride{1};
 
     VolumeOverlayController* _volumeOverlay{nullptr};
+    SurfacePatchIndex _surfacePatchIndex;
+    bool _surfacePatchIndexDirty{true};
+    std::unordered_map<const QuadSurface*, int> _surfaceDirtyBoundsVersions;
+    std::unordered_set<QuadSurface*> _indexedSurfaces;
+    std::vector<QuadSurface*> _pendingSurfacePatchIndexSurfaces;
+    QFutureWatcher<std::shared_ptr<SurfacePatchIndex>>* _surfacePatchIndexWatcher{nullptr};
+
+    void rebuildSurfacePatchIndexIfNeeded();
 };
